Replace unmaintained `react-document-title` with `react-side-effect`
`react-document-title` has not been updated since 2017.  It declares a dependency
on a similarly outdated version of `react-side-effect`, which makes use of a React
API that is now deprecated (`componentWillMount`), which produces warnings in the
console.

Remove our dependency on react-document-title, and then port the DocumentTitle
example from https://www.npmjs.com/package/react-side-effect to Typescript.

Fixes #334