Reject calls to `insert()` that have extra field not declared in schema
f66ec97c375f46717def29a5a4d653ff178540b8 fixed a bug in which I failed to add a
field to the schema, which caused the field to be dropped when serialized to
the database, which caused garbage collection of `CallParticipants` to do
nothing.

I expected `insert()` to reject types with fields that were not known to the
schema, but Typescript is generally permissive.

With this change, we reject arguments with unknown fields.  It catches the bug
fixed above via the typechecker.