Setup STUN/TURN server on app instances
In order for this to work, we need to be able to connect direct to the
instances, not through the ALB (which can't handle UDP and doesn't
guarantee persistent mappings). Therefore, this change also adds a
Lambda which monitors for changes in the ASG membership and updates a
DNS multivalue record to point to all current members of the ASG. This
is usable as the address for the STUN/TURN server.