Close popover when clicking into iframe
Previously, on devices that do not support hover, clicking into an iframe did not dismiss the popover. Per @zarvox's suggestion, solve this by adding a blur event handler to the window. This has the side effect of closing popovers when switching tabs (on mobile - elsewhere they were already closed by mouseLeave).

This fix requires switching from focus to click as the secondary trigger. As a result, clicking the tag can also close the popover (on any platform), which seems reasonable.