Only one server opens a Discord gateway connection
When we get a new bot credential, create a discord.js client, but only
open the websocket connection if we manage to grab a lock. As long as we
can keep the connection healthy, hold onto the lock. If the connection
becomes unhealthy, release the lock.

This attempts to also release the lock on process shutdown, but my
testing has shown that isn't especially reliable (at least on Windows,
it might be better on Linux). In any case, even if it takes 10 seconds
for the lock preemption to trigger, Discord seems to still allow posting
messages without an _active_ gateway connection, so long as there was a
_recent_ gateway connection, so as long as we recover relatively
quickly we should be fine.