Block until fully connected to Discord
I thought that `client.login` returned a promise that wasn't fulfilled until the
client was fully connected, but that doesn't seem to be the case. I observed
circumstances where the connection status was `IDLE` immediately after the
promise resolved, which would cause the connection management loop to exit
immediately, leaving nobody holding the gateway connection.

Instead use the ready event as a trigger.