Soft-delete CallParticipants
It'll be useful to have some data with which to gather statistics about how RTC
calls were used after Hunt ends, so we can understand usage patterns and optimize
for the ways people actually seem to work.

We don't intend to soft-delete CallSignals at this time -- they're more
verbose, fill a much more transient role (temporary IP/port combos, including
on our relay server), and don't add a ton of semantic information that we
think would be useful for understanding usage right now.  (Though it would
certainly be interesting to know what fraction of clients wind up using the
relay or not!)