Make initial page load faster for operators
Historically, in order to display pending guesses, the operator view has
loaded all hunts and puzzles. If a Jolly Roger instance has a lot of
history, that can be quite expensive. Instead, expose a special
subscription that includes pending guesses + the specific hunts/puzzles
that they're linked to. This requires more state tracking on the server,
but since there are relatively few operators it seems worth that cost.