Synthesize new guess submissions in puzzle chat
This will help make it more visible when someone submits a new guess,
so you don't get multiple people submitting the same guess as often.
Most users aren't operators, so the only hint that someone else has
submitted a guess is the number on the `Guesses (1)` button changing,
which isn't super noticeable.

Fixes #290.