Enable hot-module-replacement
Hot module replacement allows Meteor to do pseudo-refreshes of the page,
updating module bindings, without actually having to do a full refresh cycle.
This allows iterating on UI elements with a round-trip time from file save to
"change visible in browser" in as little as one second (e.g. changing some text
in an <h1> in a React component).

See https://docs.meteor.com/packages/hot-module-replacement.html for more details,
but the upshot of this change is roughly "the edit-compile-test loop will get
faster some of the time."