PuzzleListPage search: show solved if no unsolved matches found
Later on in hunt, it's often the case that you have a bazillion solved
puzzles and like 10 unsolved ones, and you'd like to toggle that "show
solved" button.  But then in another moment you're trying to find the
link to the puzzle that gave you a particular answer or something, and
you can't quite remember the name or if you've solved it or not, and so
you have to toggle show solved again, and then you have to toggle it
back again after you're done searching.

That is a drag.  If your search would return no matches, it's more
useful to show *something* that matches the search string but not the
"show solved" constraint, if such a thing exists.

So this patchset does that -- if filtering by search term gives nonzero
results, but additionally filtering by solve state returns no results,
we'll show the matches that are solved, along with message box
explaining the situation.

Fixes #424.