Save last "Show solved" state per Hunt in localStorage
And use that value as the default value on PuzzleListPage for that Hunt.
This has the effect of making the toggle "sticky", which makes the "hide
solved" view more useful, especially later on in hunts when there are
many solved puzzles and we're trying to see the ones that remain more
clearly.

Now that we have UI that indicates that we're hiding some subset of
puzzles when applying either a search or an unsolved filter, I'm less
concerned about folks accidentally hiding solved puzzles and being
confused when they can't find the puzzle they're looking for, so I'm
more comfortable with the idea of hiding solved puzzles more of the
time.

Fixes #423.