Make team name configurable
We add one more `Setting` value to wrap up a string containing the
server name.  We plumb it to the client via a simple pseudocollection.
(Since there's only one such value, it didn't seem worth resurrecting
`PublicConfig`.)

The team name value is used in ~three places:

* in client code, on the login page
* in server code, for puzzle Google doc/sheet titles (two usages)
* in client code, in the user help text talking about the $TEAMNAME
  Discord server.

This change completes our work on custom branding.  With this patchset,
it is now possible to operate a Jolly Roger instance for a team that is
not Death and Mayhem, using the exact same code as upstream, which
should help encourage any teams that wish to operate a Jolly Roger
instance of their own to develop and use the upstream builds, rather
than forking.

As a bonus, it means that our "operating a Jolly Roger instance"
instructions could potentially just tell people to use our public Docker
images directly and bring their own `MONGO_URL`, `MONGO_OPLOG_URL` and
`MAIL_URL`.

Fixes #193.