Make breadcrumb backing implementation hook-based
There's a bunch of subtlety in this (just as there was a bunch in the
previous breadcrumbs implementation) but this appears to produce the
desired visual result.

Adds a backwards-compatible `withBreadcrumb` higher-order component
which makes use of the newly-introduced `useBreadcrumb`.

This works, but since we're calling setState in the middle of another
component's render function, React complains.  We'll fix that in a
follow-up commit.  In the meantime, this is a stable intermediate state
for migration.