Add async loading for SetupPage and RTCDebugPage
This is made pretty straightforward with the introduction of React.lazy
and the Suspense component. The only thing we really need to provide is
a placeholder component to use as a loading indicator.

This doesn't seem to make much of a dent in the bundle size, likely
because most of the bundle is dependencies, and I don't think removing
these pages allows for pruning many dependencies, but it still seems
like good hygiene and a good leaping off point for more code splitting
later.