Drop moment.js dependency
We used a very narrow subset of moment.js in client-side code which
probably doesn't justify the 60 KB or so that it adds to bundle size.
Instead, replace it with a hand-tooled function that replicates similar
formatting rules.

On the server side, we only used it for some pretty simple second-based
arithmetic, which is pretty trivial to replicate with native JavaScript.

(Note that this isn't identical because I realized on inspection that
our old handling of dates in the past 7 days vs. the future 7 days was a
little ambiguous. I don't think this comes up super often since I don't
think we show a lot of future-dated things, but this seems less
ambiguous so a bit better.)