Rework how we patch our schema into Meteor.users
Previously, we required imports/lib/models/users for its side-effects.
This had the unfortunate impact of requiring the entire set of schema
libraries just to run the app.

Instead, have meteor_users re-export the `Meteor.users` object, and
switch to using that anywhere that we need to get at Meteor.users. This
loads the schema attachment on-demand, instead of by default.