Make PendingAnnouncements.remove use a Meteor method
This is the last direct client-side collection modification.  With this
removed, we no longer rely on the client being able to directly modify
collection data, and can later remove the collection allow/deny rules.

(We might add back in a rule allowing admins to do anything all the
time, but that seems appropriately orthogonal.)