Convert AccountForm
There were a bunch of unused styles in at_bootstrap.css, including some
that seem like they were _intended_ to be used (e.g. `at-pwd-form`,
which is a class in the CSS but an id in the markup). I chose to not
convert those since we survived without them before. I also chose to
drop the stylings for `.at-btn` and `.at-form p` because they were very
minor style tweaks.