Allow specifying {created,updated}{At,By} fields
This was arguably a security concern when we were exposing the Mongo
write methods directly to clients, but with @zarvox's permissions rework
in #463, all Mongo changes are mediated by the server, so this should be
safe.

It also allows us to use the Base class in cases where we create records
in background processes that aren't directly connected to a client-side
call.