Upgrade @types/meteor
The newer version of @types/meteor is a bit more sensitive about null
vs. undefined. All of the new sensitivity is in places that don't matter
to us (i.e. we were passing in nullable values that we knew couldn't be
null), so just rework the interfaces to make it clear that things are
not null.

The types also include their own declarations of
Promise.{await,awaitAll,async,asyncApply,prototype.await} so we no
longer need to provide our own declaration.