Fix Discord object selection
There are a couple of scattered bugs here which mostly come down to
Meteor's check function being strict about object keys and Typescript
being loose. In Meteor, it's an error to pass keys in an object that
aren't part of the checked shape. In Typescript, this is allowed (the
keys just aren't accessible).

We were silently ferrying around all of the keys from the Discord cache
record, which caused an error at check time. Instead, pick off the
specific keys we care about.

(Also fix a small typo in the setup page server logic)

Fixes #470.