Make Server model a bit more useful
Move the Server model to a shared model, instead of a server-only model.
Since so much of the RTC state is tied to the server, it's going to be
important to have access to it from the client. Additionally, capture a
PID and hostname for the server process, so that if we need to debug
something we know where to go.