Reimplement audio calls using a Mediasoup SFU
Instead of using peer-to-peer WebRTC connections, run a Mediasoup SFU on
the server and connect to it using Mediasoup's client library. On the
server, each active call corresponds to a Mediasoup router, which is
created on the server that is handling requests for the first peer to
connect.

Other peers might have a Meteor connection to a different server, which
implies that all communication needs to happen via the database. The
server and client communicate with each other by creating various
database records.

For ease of understanding, this replaces the CallParticiant and
CallSignal models from the original implementation, and breaks out the
client-server communication into many different models. (In theory, I
think that most could have been crammed into CallSignals, but this was
easier to reason about.)

This gets us down to 2 connections from a given peer - one for sending
and one for receiving (the Mediasoup client library only supports
uni-directional connections). Structurally, on the client, this means
that CallSection is now responsible for connecting to the SFU (instead
of each instance of CallLinkBox, which no longer exists, being
responsible), and fanning out components for each peer. (A client
receives a consumer with an audio track for each peer.)

Using a central SFU server that is known to be internet accessible means
that we no longer require STUN/TURN for peer-to-peer NAT busting, so
that configuration is all removed.

Additionally, this clamps down the timeouts on dead server garbage
collection - since all call traffic goes through the server, a dead
server is potentially a much larger impediment to collaboration, so
liveness matters more than it used to.