Switch to host networking in CloudFormation config
Now that we're running RTC traffic directly through the jolly-roger
container, we should move it to host networking to save on overhead. The
easiest way to make this work was to just switch to the standard nginx
container and provide our own nginx config (which is mostly taken from
the config that nginx-proxy was previously rendering).