Make debugging in prod slightly more pleasant
Switching from Meteor's standard minifier to the zodern packages emits
sourcemaps. The add-react-displayname Babel plugin adds a `displayName`
property to React components, which causes them to show up with
unmangled names in the React dev tools (which apparently can't use
sourcemaps to pierce the minification on its own).

Neither of these are perfect - in particular, some of the libraries we
include seem to not get included in sourcemaps or have their display
names tagged correctly - but I think this is still a meaningful
improvement.