Rototill the build to be entirely Docker driven
Our build configuration was originally written before multi-stage Docker
builds and for Docker Hub. Since Docker Hub had no caching of
intermediate images and we wanted to generate the smallest images
possible, there was no advantage to writing a Dockerfile that was
cacheable.

Now that we have a native dependency with a heavyweight build, we would
benefit significantly from a build that could cache intermediate steps,
and with GitHub Actions and multi-stage builds, we can do that.

Switch to the docker/build-push-action, which among other things,
supports the GitHub Actions Cache API, and use that for both test and
build. Additionally, switch to the GitHub Container Registry, which
allows us to eliminate a shared secret for authentication (since GitHub
provides an authentication token as part of the GitHub Actions
workflow). Finally, enabling the new docker/setup-buildx-action means
that we can lightly parallelize the two stages of the build.

Fixes #494.