Enable nginx buffering
The nginx config was already sufficient to cache any response that the
application claims is cachable (by way of the Cache-Control header). But
nginx requires buffering to be enabled in order to cache any responses.

(The internet led me to be slightly concerned that this would cause
problems with sockjs, but (a) websockets seem to work fine (b) we don't
support any browsers that don't support websockets (c) the sockjs
protocol spec specifically says that XHR URLs must disable caching)