Add FirehosePage for tailing/searching all chat messages
This page adds an in-app view of the sort that we historically aimed for
from the firehose in IRC, Slack, or Discord, where we could see all chat
messages on all puzzles in a single interleaved time-series view.

I'm ignoring formatting here for now; this is largely to make searching
over the whole of a hunt's chatlogs easier.

The filter search only looks at the message text; if you wanted to
filter on a single puzzle, just go view the log from the puzzle page.