Rework chat scroll position behavior
Most of the time, users are interested in the most recent chunk of chat
history, and think of the backlog as being pinned to the bottom, rather
than the top.  However, in the web APIs, overflow definitely defaults to
considering space from the top-left corner of a rectangle.

To approximate the "fixed distance from bottom of scrollback" behavior,
we save the distance from the bottom that we're aiming for, and have a
function which jumps to the appropriate scrollTop value to maintain that
distance-from-bottom which we call in response to a number of possible
ways that might cause the space available to the chat history pane (or
the space demanded by the contents to be displayed) to change:

* Callers or viewers are collapsed or expanded
* New callers or viewers join or leave, causing their boxes to take up a
  different number of rows
* Other transient RTC warnings are displayed
* The sidebar is resized (via the SplitPanePlus drag)
* The browser window as a whole is resized
* The chat input box grows or shrinks due to multiline input
* A new chat message arrives

There's one noteworthy exception: if you scroll partway back in the chat
backlog, you probably don't actually want the chat jumping around as the
scroll height of the chat history grows in response to new messages
coming in, since that makes it hard to read.  So when the message log
changes, if we were scrolled away from the bottom, we instead update the
target distance from bottom.  (When we send a chat message, we do wish
to show it in the backscroll to avoid user confusion, so we jump to the
bottom.)

We were missing a number of possible "size available has changed" event
sources -- nearly every render change in ChatPeople can cause its height
to change, and we also failed to keep the history pinned on sidebar
resize.  This change attempts to rectify these omissions.  Because we
operate on refs and wish to make DOM mutations, we should make use of
`useLayoutEffect` rather than `useEffect` anywhere we'd want to react
before the browser makes the changes visible to the user, since
otherwise we'd see flicker as `scrollTop` jumps around between frames.

Fixes #427.
Fixes #428.