Update eslint + airbnb profile and fix new errors
Virtually all of these are the result of more strict settings for
`function-paren-newline`, `function-call-argument-newline`, and `semi`
and could be autofixed. The new default enables
`react/function-component-definition`, but this changes the config to
prefer arrow functions, which we use almost exclusively.

The small number of manual fixes involve a few function
declaration-style components that were converted to arrow functions, a
useless React fragment (with a single child), and one case of
`no-promise-executor-return` in the tests.