Use our own breadcrumb implementation
Upgrading Bootstrap caused a bug because they began floating the
separator. This works if you're using display: flex, but we were
overriding with display: block to get better wrapping behavior. But
given that we were overriding Bootstrap's styling so heavily, we weren't
getting any benefit from using the Bootstrap classes in the first place,
so just switch to our own version.