Rework acceptance tests
React Router v6 doesn't support creating your own memory history and
passing it in, so we need a different way to inspect state. We can do
that by creating a component that threads navigation back up to the
parent.

We can also use this as an opportunity to drop enzyme, since React has
its own native helpers for testing now.