Fix tests in CI build
When I split the build and test steps into separate Docker stages, I
inadvertently bypassed running the tests at all, because nothing from
the test stage was copied into the final production stage. Docker was
smart enough to skip those steps entirely.

I want to force them to run without un-parallelizing the test + build
steps, so make sure we use the output of the test step by copying the
scripts directory out of it. It's a bit of a roundabout way to do
things, since we don't actually change the scripts directory at all in
that stage, but it creates the dependency we care about.

This also requires some additional patches to make the browser tests
actually work, including installing dependencies that aren't included
with puppeteer's copy of Chromium.