Narrow role/channel listings to selected guild
If the Jolly Roger Discord bot has been added to more than one guild,
channels and roles from all of those guilds will be available in the
Discord cache for selection when creating/editing a hunt, which is not
the behavior we want.

This doesn't come up often in practice because generally Jolly Roger
runs in a single-guild setup, but this gives us future proofing (and
also lets us slightly simplify our role logic).