Replace `Meteor.subscribe` with `useSubscribe`
In any React components where we were creating subscriptions with
`Meteor.subscribe` inside of a `useTracker`, replace that with the new
`useSubscribe` hook. This prevents spurious churning of subscriptions -
`useTracker` reruns the reactive function whenever any of its
dependencies change, and that causes any subscriptions in the reactive
function to be re-created and torn down. By using the separate
`useSubscribe` hook, we avoid that unnecessary churn.

(This is in practice not user-visible, except for subscriptions that
cause side-effects like the RTC or viewer count subscriptions, but it
does cause churn in the client-server communication that we can avoid).

This does leave a few calls to `Meteor.subscribe` - namely those that
are done at the top-level for data that should always be present on the
client.