Destructure values returned from useTracker
Instead of returning a tracker object, destructure the individual values
that are returned into top-level variables in the component. In addition
to reading a bit more fluidly, I find this to be an improvement because
it makes it harder to fetch data that you don't ultimately use,
especially when compared to a declared type for the tracker object.