Avoid passing complex objects via props
This change is basically in preparation of switching to the new
`useFind` hook at some point. Today, any time that a query result
changes, we get a new list of new objects, meaning that any comparison
of old and new query results will show unequal. This (e.g.)
significantly reduces the effectiveness of `React.memo`, which only does
a shallow comparison. `useFind` attempts to make the individual database
objects that it returns have a stable identity.

However, it doesn't guarantee any stability for the array that it
returns (and making that stable is generally difficult within the
limitations of React). This means that passing lists or dictionaries of
database objects causes more child re-rendering than might otherwise be
necessary.

To try and lightly ameliorate that, this commit tries to reduce the
number of places where we pass a list or dictionary of database objects,
either by reducing to a scalar value in the parent component (e.g.
looking up a display name rather than passing a dictionary of display
names) or by pushing the query into the child component (in cases where
that data wasn't reused by multiple children).