Track recent activity in voice chat
Use AudioLevelObserver + some debouncing to keep track of the last time
there was voice activity (with roughly 60 second granularity), and keep
that on a new CallHistories model (which persists across calls being
started/stopped, unlike any of the other Mediasoup models).

This data gets displayed in the call section if you're not currently
joined to a call.