Put Google Drive documents in per-hunt folders
Instead of dumping documents in the root of the Google Drive of the
linked account (ugh), create a separate folder for each hunt and put all
of the puzzles for that hunt in there. This includes some logic to try
and converge the structure as documents are accessed, and a button on
the setup page to force a full reorganization.

It also includes a new "root folder" setting, so that we can put the
hunt folders in a folder of their own if that's desirable.