Defer layout effect callback that we can't fire due to null refs
There's a couple places where we need height changes from other children
to trigger scrolling in another child of a component.  We had been
dispatching these events via callbacks that accessed refs, but it turns
out that under certain render batching and ordering, we would fail to
deliver the callback to the pane that needed to be scrolled.

What's more, when we miss delivering such an event, the browser
generates a scroll event on the pane to be scrolled which we must
interpret as the user scrolling the pane to that target, breaking
the previous scroll offset, and causing us to no longer keep the
chat backlog sticky at the bottom.

This patchset attempts to fix the misbehavior by adding a ref (because
we must act within *this* render batch, not later due to a setState) for
keeping track of if we need to deliver that callback once the ref is
nonnull, along with a useLayoutEffect in the parent which will check the
refs and call the callback as appropriate before yielding back to the
browser to render the changes to the DOM.