Show puzzle age & recent voice activity next to viewer count
In the interest of showing more context about each puzzle on the puzzle
list page, for unsolved puzzles, next to the subscriber count, we've added:

* time since unlock, which is nice to peruse to see what puzzles are
  older and how long they've been there
* an icon indicating time since last voice detected in the audio call

These generally aren't the most important data on the page, though, so
I've made them slightly smaller font and muted in color, so as to reduce
the space they take up and the amount they draw the eye.

Fixes #299.