Use destructuring assignment for React props
There are a few components (AccountForm and Tag) where this isn't
practical because we're using a disjunction type to capture when certain
parameters are optional or not, and there's no way to do a destructuring
assignment that works across both sides of the disjunction.

I left them as-is for now, but at some point we should maybe reevaluate
if this is the best way to configure the types.