Convert much of puzzle.scss to locally-styled components
This change untangles and flattens some pretty tricky contextual CSS
from the PuzzleListPage and PuzzlePage to try to make things read more
straightforwardly from the perspective of wherever a particular
requirement for a style comes from, rather than having lots of nested
selectors.

Of note: the `grid` and `table` RelatedPuzzle layouts wound up sharing
exceedingly little in practice, but they were rather coupled to certain
CSS classes, so I decided that they would be better off split out as
separate components so they could be changed and evolve more
independently. `RelatedPuzzleTable` and `PuzzleTable` are separate
components filling the role of the previous `table` layout used for the
overlay when rendering Tags in the metadata section of PuzzlePage.

I also took this opportunity to scrutinize and adjust the styling of the
answer-removal buttons next to each answer in "no operator queue" mode,
so they should be aligned better now.  Other than that, there shouldn't
be any visible changes resulting from this patchset.

There's still more to convert from puzzle.scss, but given that this is
already rather large and so much of the styling heavily interconnected,
I figured this was probably a good chunk to carve off for now.  Future
changes will hopefully be smaller changes and easier to review.