Use $and for query constraints instead of mutation
Previously, for publishing data, when the client sent us a query filter,
we would mutate it in order to add a constraint on (e.g.) hunt
membership. However, we can instead use Mongo's built in $and operator
to specify a separate query term that documents must satisfy. This is
much more straightforward and less error prone.

(I also specifically verified that Mongo's query planner is able to
handle this case - if, e.g., you have one $and branch that specifies
hunt: $in, and another that specifies hunt: $eq, it will merge those and
create a single set of index bounds on the $eq value.)