Merge profiles onto the Meteor.User object
We historically avoided using the profile field on Meteor.Users due to
its relatively lax default permissioning scheme. But those defaults are
not difficult to change, and having user information spread across two
different collections makes a lot of query patterns difficult (since
they effectively require joins).

So instead, migrate the old profiles model onto the Meteor.User object.
This is intended to be a fairly 1-for-1 migration in order to make
review easier, so it does not do things like add additional constraints
to data visibility relative to what we have historically done; that's
reserved as a future task.