Nuke meteor-down load test framework
Having a load test framework was a neat idea, but it's not really that
useful in practice. The harness is brittle - nothing verifies that the
test uses the same subscriptions as the production code, so it can get
(and presumably has gotten) out of sync. It's insufficient to load test
a modern jolly-roger install, since it has no knowledge of audio calls
(and adding support would be quite elaborate). And finally it just
doesn't have that much utility - it hasn't seen meaningful edits since
2017, and I can't remember the last time I actually tried to run the
code.

We can always retrieve this code from the git history if we find it
useful later, but in practice I think we're better off not having this
polluting current revisions of the code.