Move profile fields to top-level of Meteor.User
Meteor's support for merging different fields on the same object across
multiple subscriptions doesn't work for nested objects, meaning that a
subscription with `fields: {'profile.displayName': 1}` and a
subscription with `fields: {'profile.discordAccount': 1}` will clash,
only returning one of the two fields.

Given that this is a somewhat fundamental restriction on the current
Meteor DDP protocol, the easiest solution is to promote the fields from
the user profile to top-level themselves.