Add a new smoke test to verify that components render
This is pretty dang basic, but gives us a bare minimum validation that
no route is totally broken. I verified that adding a "throw Error" to
the top of (say) the AnnouncementsPage component causes this to fail
(since even though the exception gets caught by our error boundary, it
still shows up as an uncaught exception, which mocha picks up).