Tweak puzzle activity monitor
Increased the spacing between activity monitor items a little bit, and
adjusted the column width so it wouldn't overflow.

Moved the call activity icon to the third position, and hid it if
there's never been a call.

The icon is after the text so that it lines up in a column with all the other
icons of its ilk. The text is right-aligned so that it's easy to scan down
and see the # of digits of people in a puzzle.

On mobile, the activity monitor moves down into its own row, between the puzzle
name and the puzzle link.