Push display names as part of global subscriptions
For subscriptions that are not scoped to a hunt (which, conveniently, is
also the subscriptions that use JoinPublish + the rtcdebug page, which
already has a custom subscription), push the display names that we need
as well. This lets us get rid of `useSubscribeDisplayNames` in any place
where we don't already have a hunt-level scope.