Add SwappableCursorPublisher
Publishes in Meteor are not reactive, but we occasionally want to be
able to swap out the query that is being published without tearing down
and recreating the subscription.

This mostly just wires observeChanges up to the added/changed/removed
methods on Subscription, but with a little extra magic to handle
swapping out one cursor for another.