Make user/profile subscriptions reactive + permissioned
This makes a handful of changes to the various publications for users
and profiles. First, it makes them all reactive to changes in hunt
membership of the user making the subscription. This is necessary so
that (e.g.) the all profile list page updates when the user is joined to
a new hunt. It's probably not necessary to do this for all of these
publishes, but it's easier to do all of them than figure out which ones
need it.

This also separates out the old `profiles` subscription into
`allProfiles` and `huntProfiles`, and drops `userInfo` which was no
longer used now that role information was only shown on the hunt profile
list page.

Finally, it limits publications to only publish users with whom you have
a hunt in common. For `allProfiles`, `huntProfiles`, and `profile`, this
includes filtering the user's list of hunts to hunts you are also a
member of. And for `huntRoles`, this only publishes roles for a specific
hunt, rather than roles for all hunts.

Fixes #629.