Store all puzzle list state in localStorage
This includes the collapse/uncollapsed state of individual groups as
well as the group/unlock preference. In order to prevent proliferation
of localStorage keys, everything is switched to use-persisted-state and
grouped under a single key as a data structure.

To avoid confusion when searching, we ignore the collapse state of all
groups when showing any search results. And in order to avoid needing to
do a bunch of clicks, there's a new "expand all" button that will expand
any collapsed groups.