Improve fixture hunt data
This replaces the old fixture data from the 2015 hunt with a new set of
fixtures from the 2018 hunt. This improves on the old fixtures in a few
notable ways:

- The hunt is no longer complete. Our old fixture data had the hunt
  after all puzzles had been solved. The new fixture data is about
  halfway through the hunt (the team in our fixtures has fully solved
  the Emotions round, unlocked Hacking Island and Games Island but not
  completely solved either). We also now include guesses. The old
  fixtures didn't create guess objects, meaning that the guess queue
  didn't work as expected.
- The hunt has interesting structural components. The 2015 hunt was a
  great hunt, but mostly flat structurally. Thus it provides a
  suboptimal example of Jolly Roger's more sophisticated organization
  schemes. 2018 had much more going on structurally (including
  overlapping metas and multi-level round structure)

In order to make the experience more pleasant, if you are an admin and
there are no hunts, we add a button to the hunt list page to create the
fixtures (instead of needing to invoke the Meteor API method manually
from the JS console).