Move hunt creation/edit to a separate page
As we add more hunt configuration, we're rapidly outgrowing what's
reasonable to put into a relatively small modal. Given that, push it to
a separate page.

This is mostly pure copy-paste of the old modal content, but with a few
changes - namely handling breadcrumbs and URL parameters (instead of
having a hunt passed in), and slight changes to the post-submit flow.
Instead of closing the modal, we take different behavior depending on
whether we're creating a new hunt or editing an existing one. For
creation, we redirect to the newly created hunt. For edits, we scroll to
the bottom of the page and display a banner confirming that the edit was
successful, with a back button to take you back to the hunt list.