Rework forms to limit width and put buttons right
Save buttons feel like they should always be at the bottom right of the
form they're modifying.  To keep them from getting lost, we should limit
the max-width of forms.

Large amounts of text (like that on the SetupPage) is also easier to
read when slightly narrower, and there's less of an argument for
information density on that page vs the PuzzleListPage, so I limited the
width of the content of each section..

I also took the opportunity to slightly tweak the design on SetupPage,
making the section headers fill the full width, making the toggle button
pairs into a reusable component, and adjusting some padding to avoid
buttons right on the edge of their containing element.